Add another URL parameter to ease detection of imports from exporter
It doesn't look like there is a way to make this a POST request so
a GET will have to do for now.

This does mean that on the other side of the system people may be
able to copy URLs once clicking the export button and then trick us
by giving it to other or reloading the page.

Bug: T187538
Change-Id: I89a157f5239bf5e5dddb2f8f4c41d41e8bf41ce5