Hide dysfunctional Beta feature for non-autoconfirmed users
This is fixing an inconsistency: For non-autoconfirmed users enabling
the Beta feature does not do anything. It's a dead button. Let's remove
it please.

We will discuss this further in one of the next story times. But
independend from that I suggest to first fix this inconsistency. We can
later change to an other consistent solution, if we find one.

Bug: T160891
Change-Id: I835e4e80cbb7cc6d850445998c10cde2b811898f