More broad check for test wikis
I found the two string comparisons to specific, and wanted the code to
make more sense: The message "Export to Test Wiki" is intentionally so
generic that it covers any "test wiki". The comparison should reflect
this. With this change it also allows "test1.…", "test3.…" and so on,
getting rid of the limitation to "test." and "test2." only.

Feel free to disagree if you think this is going the wrong direction.

Change-Id: I0c14f5b9271a9b1925af2c1d3601c3964c7d9e00