Streamline export URL query parameter building
I found a bit of duplicate code here and feel this can be avoided,
hopefully making this code a little easier to understand. It's also
much more flexible with this, as this allows to collect any number of
URL parameters. There are only 2 at the moment, but even for these the
code already became a bit complicated.

Change-Id: I66ec8905322113e7f8a0e9070caaa5cf5240f1e2