Avoid using User::getOption
Remove using of User:getOption since this method
will be hard-deprecated. User::getOption was mocked
in FileExporterHooksTest::testOnSkinTemplateNavigation_betaDisabled
to call it in BetaFeatures::isFeatureEnabled of
BetaFeatures extension. Since the use of User::getOption
in that method was replaced with
UserOptionsLookup::getOption, we can drop the mocking
of User::getOption in FileExporterHooksTest::
testOnSkinTemplateNavigation_betaDisabled.

Bug: T296083
Change-Id: I70a2c532c9a1bd4b04a1161fabcb1309839a713e